出现内存溢出错误

+++ Code/Algorithm/IMU.h
@@ -9,11 +9,12 @@ class cIMU :public cICM42688,public PID_Inc_C
 	public:
 	float	Q[4]={1,0,0,0};
 	
-	int16_t GyroCal[3]={0};
-	int16_t AccelCal[3]={0};
+	/*修正值为浮点,乘好LSB*/
+	float GyroCal[3]={0};
+	float AccelCal[3]={0};
 
-	int16_t GyroCorrected[3]={0};
-	int16_t AccelCorrected[3]={0};
+	float GyroCorrected[3]={0};
+	float AccelCorrected[3]={0};
 	
 	void PID_Init(void)
 	{
